[Deployments] Batch relayer V4 preparation.
- Binaries, scripts and fork test.
- Deprecating V3.
- Updating README.
