--- pkg/deployments/tasks/20220720-batch-relayer-v3/output/arbitrum.json
@@ -1,5 +1,0 @@
-{
-  "BatchRelayerLibrary": "0x4f4269F7008430860693aeDB86b0C55962D6E9b7",
-  "timestamp": 1658348087919,
-  "BalancerRelayer": "0x42E49B48573c725ee32d2579060Ed06894f97002"
-}
