--- pkg/deployments/tasks/20220720-batch-relayer-v3/output/goerli.json
@@ -1,5 +1,0 @@
-{
-  "BatchRelayerLibrary": "0xdcdbf71A870cc60C6F9B621E28a7D3Ffd6Dd4965",
-  "timestamp": 1658347230144,
-  "BalancerRelayer": "0x7b9B6f094DC2Bd1c12024b0D9CC63d6993Be1888"
-}
