--- pkg/deployments/tasks/20220720-batch-relayer-v3/output/optimism.json
@@ -1,5 +1,0 @@
-{
-  "BatchRelayerLibrary": "0x5c89d9E1A668d0B39F4EC242F280bdDd445C633c",
-  "timestamp": 1658347394816,
-  "BalancerRelayer": "0x195CcCBE464EF9073d1f7A1ba1C9Bf0f56dfFFff"
-}
