--- pkg/deployments/tasks/20220720-batch-relayer-v3/output/polygon.json
@@ -1,5 +1,0 @@
-{
-  "BatchRelayerLibrary": "0x5ef4c5352882b10893b70DbcaA0C000965bd23c5",
-  "timestamp": 1658347591787,
-  "BalancerRelayer": "0xcf6a66E32dCa0e26AcC3426b851FD8aCbF12Dac7"
-}
