+++ pkg/deployments/tasks/20220906-batch-relayer-v4/readme.md
@@ -1,0 +1,10 @@
+# 2022-09-06 - Batch Relayer V4
+
+Deployment of the fourth `BalancerRelayer` using `BatchRelayerLibrary`, for combining multiple operations (swaps, joins, etc.) in a single transaction.
+
+This new version adds support for Liquidity Gauge interactions, including staking, unstaking, minting, etc.
+
+## Useful Files
+
+- [`BalancerRelayer` ABI](./abi/BalancerRelayer.json)
+- [`BatchRelayerLibrary` ABI](./abi/BatchRelayerLibrary.json)
