+++ pkg/deployments/tasks/deprecated/20220720-batch-relayer-v3/output/mainnet.json
@@ -1,0 +1,5 @@
+{
+  "BatchRelayerLibrary": "0xD966d712F470067B60D37246404D6DFe5Bf0B419",
+  "timestamp": 1658346757236,
+  "BalancerRelayer": "0x886A3Ec7bcC508B8795990B60Fa21f85F9dB7948"
+}
